+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### CarPlay
 
+* Added new CarPlay icons.
 * Added logic to automatically recenter puck after canceling preview mode.
 * Fixed Car Play route does not properly cancel when navigating solely from the Car Play application.